Activities directory relocation and code cleanup
All the activities were sent to a folder in order to improve the project's structure.
Added comments on the overall functionality of the RecommendationFragment.
Removed unnecessary logs from the CameraFragment.
HEX number used to remove alpha values from colors was turned into a variable on the CameraFragment and the ColorFormActivity.
Recommendation card had the size of all the TextViews reduced to have more evenly sized cards as cards with more text became larger.
RecommendationFragment's CardStack changed from showing 4 cards to 2 in order to keep UI cleaner.