RecommendationFragment updates
RecommendationFragment now holds to buttons at the bottom of the CardStack to press when the users runs out of recommendations.
The buttons add functionality to take a new photo for recommendations or fetch more songs, this second option is executed on the UI thread in order to be able to modify the visual elements of the fragment.
Song model modified locally and on the DB backend to include the names of the song, the artist, the album, and Spotify's ID for the song.