Login and SongHistory bugfixing
Fixed a bug where the login always sent the user to the ColorFormActivity disregarding if the user had set the colors or not before. It is now working as intended only sending new users or users with no ColorRelation form.
SongHistory showed songs from all users to a single user, it now only shows songs the user has liked.