Song preview and like bugfixing
Songs can now play a preview if available by tapping the card on the recommendation stack. This preview is also played automatically on the SongDetailFragment.
Fixed a bug that occurred when a song was liked, said song was not pushed to Parse, rather than the last song of the stack. This is now working as intended pushing the swiped song.