Code cleanup on HEX extraction and UserProfileFragment visual update
Extraction and setting of colors for the ColorFormActivity and CameraFragment has been updated to have less repeated code and increase readability.
RecommendationFragment got cleaned up
UserProfileFragment has been modified to include the current color relation the user with a similar format to the ColorFormActivity.
Also included 2 new vectors to use in the UserProfileFragment, a filled circle and a round user icon.